requirements.txt
pdfplumber==0.5.23
tabula-py==2.1.1
pandas~=1.2.4
